refactor: use fixture filename in todo example for footer and header
