+++ examples/todo/src/components/Footer/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { TodoFilter } from '../types.js';
-import { useTodoContext } from './TodoContext.js';
+import { TodoFilter } from '../../types.js';
+import { useTodoContext } from '../TodoContext.js';
 
 export function Footer() {
   const { todos, setTodos } = useTodoContext();
